+++ src/main/java/com/gp/beershop/repository/CustomerOrderRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CustomerOrderRepository extends JpaRepository<CustomerOrderEntity, Integer> {
+public interface CustomerOrderRepository extends JpaRepository<CustomerOrderEntity, Long> {
 
 }
